incorporate @Jimbilben's power analysis work into Bookdown
@Jimbilben  I incorporated this as an 'appendix', but we might ultimately move (part of?) it into the power analysis section

I've read, commented on, and made some adjustments (look on these as suggestions, we can revert back) on the first half. Now I'm hoping to dig into the 'concrete' (code) example.

If you open any of the doc/*.html files (in a browser) you can see how it looks. I think the 'cold folding' is a pretty big advantage over the pdf