1. Added 'hardlinked' method content from EA Market testing Gitbook
@DMoss123 @Jimbilben @WillemSleegers

In the 'EA market testing Gitbook' (which is written in a markdown syntax), we're putting together resources and guidelines on 'how to run experiments, trials, and studies' in this context.

I divided this into sections 1. Qualitative, 2. Quantitative, and 3. Design issues ...

I also drafted a 'reporting template' for people running these studies to keep the others abreast.

But IMO this very much overlaps what *we at RP survey group* want to do in putting together guidance for ourselves and partners ... at least the less technical part.

So, I've moved this into our repo as a 'hardlink' ... edit it in one place (on my computer... or if I pull down a change from someone else) and it will be edited in the other place as well.

I *haven't* yet integrated it into our Bookdown ... we'll have to think about if and how to do that. ... maybe through a simple external link, if the markdown formats can't be easily reconciled.