lots of restructuring and cleanup ... embed web pages, etc.
@Jimbilben @WillemSleegers

I cleaned and organized this quite a bit. Willem, it might make sense for you to 'take over' the 'survey design' and 'factor analysis' sections

(not that you have to 'do' anything ... but if you are doing and writing stuff anyways, it might as well be there, I guess)

Note esp some work in the 'presentation, labeling, and vis' section. Not saying 'my way is right' ... just trying to get the ball rolling on a consensus/protocol/template on this stuff, to save us time and brainpower later